int to long

+++ FizzBuzz2/Fizzbuzz.Application/FizzBuzzService.cs
@@ -5,7 +5,7 @@ namespace FizzBuzz.Application
 {
     public class FizzBuzzService : IFizzBuzzService
     {
-        public FizzbuzzResult GenerateFizzBuzz(int start, int end)
+        public FizzbuzzResult GenerateFizzBuzz(long start, long end)
         {
             return FizzBuzzCalc.GenerateFizzBuzz(start, end);
         }
